requirements.txt
flask==2.0.2
flask-sock==0.5.2
APScheduler==3.9.1
python-dateutil==2.8.2
profanity==1.1